SKALE-1857 Refactor ssl_utils
