+++ sgx/constants.py
@@ -7,3 +7,4 @@ GENERATE_SCRIPT = 'generate.sh'
 GENERATE_SCRIPT_PATH = os.path.join(SCRIPTS_DIR_PATH, GENERATE_SCRIPT)
 
 DEFAULT_TIMEOUT = 10
+CERT_PROVIDER_PORT = '1027'
